Задание переделано
